+++ Sources/PouchFramework/Defaults.swift
@@ -1,6 +1,9 @@
+import Foundation
+
 public enum Defaults {
     public static let encryption = Cipher.xor
     public static let input = Input.environmentVariable
+    public static let representation = OutputRepresentation.staticVariables
 }
 
 public extension Defaults {