+++ Tests/PouchTests/Parsers/ConfigurationParsingTests.swift
@@ -23,7 +23,7 @@ outputs:
         XCTAssertNil(parsedConfiguration)
     }
     
-    func test_hasAtLeastOneSecretAndOneOutput_parsesSuccessfully() {
+    func test_hasAtLeastOneSecretAndOneOutput_parsesSuccessfully() throws {
         let config =
 """
 secrets:
@@ -31,14 +31,14 @@ secrets:
 outputs:
 - ./Secrets.swift
 """
-        let parsedConfiguration = try? YAMLDecoder().decode(Configuration.self, from: Data(config.utf8))
+        let parsedConfiguration = try YAMLDecoder().decode(Configuration.self, from: Data(config.utf8))
         let expectedConfiguration = Configuration(
             input: Defaults.input,
             secrets: [
                 .init(name: "API_KEY", encryption: Defaults.encryption)
             ],
             outputs: [
-                .init(decryptionFile: .init(filePath: "./Secrets.swift"), outputLanguage: .swift(.init(typeName: Defaults.Swift.typeName)))
+                .init(decryptionFile: .init(filePath: "./Secrets.swift"), representation: .staticVariables, outputLanguage: .swift(.init(typeName: Defaults.Swift.typeName)))
             ])
         XCTAssertEqual(parsedConfiguration, expectedConfiguration)
     }